issue3 refactor main (only runs app)
Deleted not used dependency from docker
