+++ docker-compose.yml
@@ -1,19 +1,9 @@
 version: '3'
 services:
-  mongodb:
-    image: mongo
-    ports:
-      - "27017:27017"
-    volumes:
-      - "mongodata:/data/db"
-    networks:
-      - network1
   api:
     build: .
     ports:
       - "8080:8000"
-    depends_on:
-      - mongodb
     deploy:
       resources:
         limits:
